custom input update with parseFloat of number type

+++ src/app/joining/_utils/custom-input.tsx
@@ -31,7 +31,9 @@ const CustomInput = ({
               onChange={(e: any) => {
                 form.setValue(
                   name,
-                  type === "number" ? parseInt(e.target.value) : e.target.value
+                  type === "number"
+                    ? parseFloat(e.target.value)
+                    : e.target.value
                 );
               }}
               defaultValue={field.value}
